+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Add `--no-checks` for `oao publish` (removes all prepublish checks),
+  useful in some cases (#62).
+
 ## 1.3.0 (2018-2-13)
 
 * Add support for extra arguments in `oao all`, e.g. `oao all ls -- -al` is